backend/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
slowapi==0.1.9
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
psutil==5.9.8
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
coverage==7.4.0